Correctly install packages.
